move tools item component
